feat: Actualizar la configuración de la base de datos y agregar la eliminación de productos por categoría y proveedor

+++ service-products/src/main/java/servicesproductos/entities/Producto.java
@@ -29,8 +29,8 @@ public class Producto {
     private int valor;
 
     @Column(name = "proveedor_idProveedor")
-    private Long providerId;
+    private int providerId;
 
     @Column(name = "categoriaId")
-    private Long categoriaId;
+    private int categoriaId;
 }
